+++ kurtosis-module/impl/services/indexer.ts
@@ -1,9 +1,10 @@
-import { EnclaveContext, ServiceID, ContainerConfig, ContainerConfigBuilder, ServiceContext, PortSpec, PortProtocol } from "kurtosis-core-api-lib";
-import log = require("loglevel");
+import { FilesArtifactUUID, EnclaveContext, ServiceID, ContainerConfig, ContainerConfigBuilder, ServiceContext, PortSpec, PortProtocol } from "kurtosis-core-api-lib";
+import * as log from "loglevel";
 import { Result, ok, err } from "neverthrow";
 import { EXEC_COMMAND_SUCCESS_EXIT_CODE } from "../consts";
 import { ContainerConfigSupplier } from "../near_module";
 import { getPrivateAndPublicUrlsForPortId, ServiceUrl } from "../service_url";
+import * as path from "path";
 
 const SERVICE_ID: ServiceID = "indexer-node"
 const IMAGE: string = "kurtosistech/near-indexer-for-explorer:7510e7f";
@@ -18,6 +19,9 @@ const GOSSIP_PUBLIC_PORT_NUM: number = 24567;
 const GOSSIP_PORT_ID = "gossip";
 const GOSSIP_PORT_SPEC = new PortSpec(GOSSIP_PORT_NUM, PortProtocol.TCP);
 const GOSSIP_PUBLIC_PORT_SPEC = new PortSpec(GOSSIP_PUBLIC_PORT_NUM, PortProtocol.TCP);
+
+const LOCALNET_CONFIG_DIRPATH_ON_MODULE = "/static-files/near-configs/localnet"
+const NEAR_CONFIGS_DIRPATH_ON_INDEXER_CONTAINER = "/root/.near"
 
 const DATABASE_URL_ENVVAR = "DATABASE_URL";
 
@@ -46,6 +50,14 @@ export async function addIndexer(
     dbName: string,
 ): Promise<Result<IndexerInfo, Error>> {
     log.info(`Adding indexer service...`);
+
+    // Send the genesis file to Kurtosis
+    const uploadLocalnetConfigResult = await enclaveCtx.uploadFiles(LOCALNET_CONFIG_DIRPATH_ON_MODULE)
+    if (uploadLocalnetConfigResult.isErr()) {
+        return err(uploadLocalnetConfigResult.error)
+    }
+    const localnetConfigFilesArtifactUuid = uploadLocalnetConfigResult.value;
+
     const usedPorts: Map<string, PortSpec> = new Map();
     usedPorts.set(RPC_PORT_ID, RPC_PORT_SPEC);
     usedPorts.set(GOSSIP_PORT_ID, GOSSIP_PORT_SPEC);
@@ -60,6 +72,9 @@ export async function addIndexer(
         `postgres://${dbUsername}:${dbUserPassword}@${dbPrivateUrl.ipAddress}:${dbPrivateUrl.portNumber}/${dbName}`
     )
 
+    const filesArtifactMounts = new Map<FilesArtifactUUID, string>();
+    filesArtifactMounts.set(localnetConfigFilesArtifactUuid, NEAR_CONFIGS_DIRPATH_ON_INDEXER_CONTAINER)
+
     const containerConfigSupplier: ContainerConfigSupplier = (ipAddr: string): Result<ContainerConfig, Error> => {
         const result: ContainerConfig = new ContainerConfigBuilder(
             IMAGE,
@@ -69,6 +84,8 @@ export async function addIndexer(
             publicPorts,
         ).withEnvironmentVariableOverrides(
             envvars
+        ).withFiles(
+            filesArtifactMounts,
         ).build();
         return ok(result);
     }